Fixed bug with non default mysql row-increments. Fixed bug in upgrade code from version 1.3.5->1.4
git-svn-id: https://plugins.svn.wordpress.org/custom-stock-widget/trunk@1114248 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ stock_widget_admin.php
@@ -116,6 +116,7 @@ function stock_widget_activate() {
         //if neither of these exist then assume initial install
         stock_widget_create_db_table();
         $values = array( //NOTE: the rest should all be the defaults
+                        'id'             => 1, //explicitly set this or else mysql configs where the default is not 1 will be broken
                         'name'           => 'Default Settings',
                         'advanced_style' => 'margin: auto;',
                         'stock_page_url' => 'https://www.google.com/finance?q=__STOCK__'
@@ -158,11 +159,17 @@ function stock_widget_handle_update() {
         case '1.3.4':
         case '1.3.5':
             stock_widget_create_db_table(); //Added table storage structure in 1.4
-
-            unset($default_settings['show_headers']); //if this exists get rid of it
-            $default_settings['name'] = 'Default Settings';
-            if (false !== sp_add_row($default_settings))
-                delete_option('stock_widget_default_settings');
+            
+            $default_settings = get_option('stock_widget_default_settings', false);
+            if ($default_settings !== false) {
+                unset($default_settings['show_headers']); //if this exists get rid of it
+                $default_settings['name'] = 'Default Settings';
+                if (false !== sp_add_row($default_settings))
+                    delete_option('stock_widget_default_settings');
+            }
+            else {
+                stock_widget_activate(); //Recall the activate function and this problem should be fixed
+            }
 
         case '1.4':
             stock_widget_create_db_table(); //bugfix for table storage in 1.4.1
